+++ src/mastodon_api/directory/types.rs
@@ -1,13 +1,13 @@
 use serde::Deserialize;
 
-fn default_page_size() -> i64 { 40 }
+fn default_page_size() -> u16 { 40 }
 
 /// https://docs.joinmastodon.org/methods/instance/directory/
 #[derive(Deserialize)]
 pub struct DirectoryQueryParams {
     #[serde(default)]
-    pub offset: i64,
+    pub offset: u16,
 
     #[serde(default = "default_page_size")]
-    pub limit: i64,
+    pub limit: u16,
 }
